// projects/frontend/src/api/TipStrategy.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Game } from "./Game";
import type { Group } from "./Group";
import type { StrategyResult } from "./StrategyResult";
import type { StrategyType } from "./StrategyType";
import type { Tip } from "./Tip";
import type { Tournament } from "./Tournament";

export type TipStrategy = {
  game: Game | null;
  tournament: Tournament | null;
  group: Group | null;
  tips: Array<Tip> | null;
  result: StrategyResult | null;
  id: string;
  tournamentId: string;
  gameId: string | null;
  groupId: string | null;
  label: string;
  description: string;
  opensAt: string | null;
  endsAt: string | null;
  calculatePointsOn: string | null;
  completed: boolean;
  strategyTypes: Array<StrategyType>;
  createdAt: string | null;
  updatedAt: string | null;
  deletedAt: string | null;
};