// projects/frontend/src/api/StrategyResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Strategy } from "./Strategy";

export type StrategyResult = {
  id: string;
  tipStrategyId: string;
  strategyResults: Array<Strategy>;
  createdAt: Date | null;
  updatedAt: Date | null;
  deletedAt: Date | null;
};